requirements.txt
h5py==2.8.0
numpy==1.15.4
Pillow==5.3.0
pkg-resources==0.0.0
six==1.11.0
